% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ky.R
\name{ky_webcams}
\alias{ky_webcams}
\title{List Jefferson County KY traffic web cameras}
\usage{
ky_webcams(limit = 100, offset = 0)
}
\arguments{
\item{limit}{Maximum results (default 100)}
\item{offset}{Result offset for pagination (default 0)}
}
\value{
tibble: id, name, status, state, county, highway, direction, description, latitude, longitude, snapshot_url
}
\description{
List Jefferson County KY traffic web cameras
}
